JetStream exception docs
